Separate example Posix things to separate module
This will prepare for adding RTOS example ports.
