+++ examples/http_example.c
@@ -41,6 +41,7 @@
 #include <sys/param.h>
 #include <errno.h>
 
+#include "u_port.h"
 #include "u_cx_at_client.h"
 #include "u_cx_at_util.h"
 #include "u_cx.h"
@@ -49,23 +50,35 @@
 #include "u_cx_socket.h"
 #include "u_cx_system.h"
 
+/* ----------------------------------------------------------------
+ * COMPILE-TIME MACROS
+ * -------------------------------------------------------------- */
+
 #define EXAMPLE_URL "www.google.com"
 
 #define URC_FLAG_NETWORK_UP         (1 << 0)
 #define URC_FLAG_SOCK_CONNECTED     (1 << 1)
 #define URC_FLAG_SOCK_DATA          (1 << 2)
 
-static uint64_t gBootTime;
-static int gUartFd;
+/* ----------------------------------------------------------------
+ * TYPES
+ * -------------------------------------------------------------- */
+
+/* ----------------------------------------------------------------
+ * STATIC PROTOTYPES
+ * -------------------------------------------------------------- */
+
+/* ----------------------------------------------------------------
+ * STATIC VARIABLES
+ * -------------------------------------------------------------- */
+
 static sem_t gUrcSem;
 static volatile uint32_t gUrcEventFlags = 0;
 uCxHandle_t ucxHandle;
 
-
-typedef struct {
-    int uartFd;
-    uCxAtClient_t *pClient;
-} uRxThreadArgs_t;
+/* ----------------------------------------------------------------
+ * STATIC FUNCTIONS
+ * -------------------------------------------------------------- */
 
 static bool waitEvent(uint32_t evtFlag, uint32_t timeout_s)
 {
@@ -90,30 +103,6 @@ static void signalEvent(uint32_t evtFlag)
     sem_post(&gUrcSem);
 }
 
-static int32_t uCxAtRead(uCxAtClient_t *pClient, void *pStreamHandle, void *pData, size_t length, int32_t timeoutMs)
-{
-    (void)pClient;
-    int uartFd = *((int *)pStreamHandle);
-
-    if (timeoutMs == 0) {
-        int available = 0;
-        ioctl(uartFd, FIONREAD, &available);
-        if (available == 0) {
-            return 0;
-        }
-    }
-
-    return (int32_t)read(uartFd, pData, length);
-}
-
-static int32_t uCxAtWrite(uCxAtClient_t *pClient, void *pStreamHandle, const void *pData, size_t length)
-{
-    (void)pClient;
-    int uartFd = *((int *)pStreamHandle);
-
-    return (int32_t)write(uartFd, pData, length);
-}
-
 static void networkUpUrc(struct uCxHandle *puCxHandle)
 {
     printf("networkUpUrc\n");
@@ -131,110 +120,13 @@ static void socketData(struct uCxHandle *puCxHandle, int32_t socket_handle, int3
     signalEvent(URC_FLAG_SOCK_DATA);
 }
 
-static int32_t getTickTimeMs(void)
-{
-    struct timespec time;
-    clock_gettime(CLOCK_MONOTONIC_RAW, &time);
-    uint64_t timeMs = (time.tv_sec * 1000) + (time.tv_nsec / (1000 * 1000));
-    return (int32_t)timeMs;
-}
-
-int32_t uPortGetTickTimeMs(void)
-{
-    int32_t timeMs = getTickTimeMs() - gBootTime;
-    return timeMs;
-}
-
-static int openUart(const char *pDevName, int baudRate, bool useFlowControl)
-{
-    int uartFd;
-    speed_t speed;
-    if (baudRate == 9600) {
-        speed = B9600;
-    } else if (baudRate == 19200) {
-        speed = B19200;
-    } else if (baudRate == 38400) {
-        speed = B38400;
-    } else if (baudRate == 57600) {
-        speed = B57600;
-    } else if (baudRate == 115200) {
-        speed = B115200;
-    } else if (baudRate == 230400) {
-        speed = B230400;
-    } else if (baudRate == 460800) {
-        speed = B460800;
-    } else if (baudRate == 921600) {
-        speed = B921600;
-    } else {
-        return -1;
-    }
-
-    uartFd = open(pDevName, O_RDWR | O_NOCTTY | O_NONBLOCK);
-    if (uartFd < 0) {
-        return uartFd;
-    }
-
-    struct termios options;
-    tcgetattr(uartFd, &options);
-    cfmakeraw(&options);
-    cfsetispeed(&options, speed);
-    cfsetospeed(&options, speed);
-
-    if (useFlowControl) {
-        options.c_cflag |= CRTSCTS;
-    } else {
-        options.c_cflag &= ~CRTSCTS;
-    }
-    // Set timed read with 100 ms timeout
-    options.c_cc[VMIN] = 1;
-    options.c_cc[VTIME] = 1;
-    if (tcsetattr(uartFd, TCSANOW, &options) != 0) {
-        return -1;
-    }
-    tcflush(uartFd, TCIOFLUSH);
-
-    return uartFd;
-}
-
-static void *readTask(void *pArg)
-{
-    uRxThreadArgs_t *pRxArgs = (uRxThreadArgs_t *)pArg;
-    struct pollfd fds[1];
-    fds[0].fd = pRxArgs->uartFd;
-    fds[0].events = POLLIN ;
-
-    while (1) {
-        int pollrc = poll(fds, 1, 100);
-        if (pollrc < 0) {
-            perror("poll");
-            break;
-        }
-        if (fds[0].revents & POLLIN) {
-            uCxAtClientHandleRx(pRxArgs->pClient);
-        }
-    }
-    return NULL;
-}
+/* ----------------------------------------------------------------
+ * PUBLIC FUNCTIONS
+ * -------------------------------------------------------------- */
 
 int main(int argc, char **argv)
 {
     uCxAtClient_t client;
-
-    static char rxBuf[1024];
-#if U_CX_USE_URC_QUEUE == 1
-    static char urcBuf[1024];
-#endif
-    static uCxAtClientConfig_t config = {
-        .pRxBuffer = &rxBuf[0],
-        .rxBufferLen = sizeof(rxBuf),
-#if U_CX_USE_URC_QUEUE == 1
-        .pUrcBuffer = &urcBuf[0],
-        .urcBufferLen = sizeof(urcBuf),
-#endif
-        .pStreamHandle = NULL,
-        .write = uCxAtWrite,
-        .read = uCxAtRead
-    };
 
     if (argc != 4) {
         fprintf(stderr, "Invalid arguments\n");
@@ -247,32 +139,10 @@ int main(int argc, char **argv)
 
     sem_init(&gUrcSem, 0, 0);
 
-    gBootTime = getTickTimeMs();
-
-    gUartFd = openUart(pDevice, 115200, true);
-    if (gUartFd < 0) {
-        printf("Failed to open UART\n");
+    uPortAtInit(&client);
+    if (!uPortAtOpen(&client, pDevice, 115200, true)) {
         return 1;
     }
-    config.pStreamHandle = &gUartFd;
-
-    pthread_t readThread;
-    uRxThreadArgs_t rxArgs = {
-        .uartFd = gUartFd,
-        .pClient = &client
-    };
-
-    pthread_attr_t attr;
-    struct sched_param param;
-    pthread_attr_init(&attr);
-    // So that the thread is tidied-up when it exits
-    pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_DETACHED);
-    pthread_attr_getschedparam(&attr, &param);
-    param.sched_priority = 9;
-    pthread_attr_setschedparam(&attr, &param);
-    pthread_create(&readThread, &attr, readTask, &rxArgs);
-
-    uCxAtClientInit(&config, &client);
 
     uCxInit(&client, &ucxHandle);
     uCxWifiRegisterStationNetworkUp(&ucxHandle, networkUpUrc);
@@ -310,6 +180,8 @@ int main(int argc, char **argv)
             printf("%s", rxData);
         }
     } while (ret > 0);
+
+    uPortAtClose(&client);
 
     sem_destroy(&gUrcSem);
 }
